I've added a transient Guid instance variable that will act as an equality test for new Notes that have not yet been committed to a database. In other words these objects are newly created and are not yet "saved".
From this I no longer need to set the ID explicitly in my service as it should not be setting an ID as it does not yet exist.

This may not be the best approach as after the Not has been saved it will be assigned an ID which will then alter the GetHashCode() result, this could actually cause a bigger problem in the end.